+++ examples/resources/pingone_role_assignment_user/import.sh
@@ -1,1 +1,1 @@
-$ terraform import pingone_role_assignment_user.example <environment_id>/<role_assignment_id>/<user_id>
+$ terraform import pingone_role_assignment_user.example <environment_id>/<user_id>/<role_assignment_id>